CXP-325: Remove DomainErrorEvent
